fix(os): unblock x86_64 SimpleOS boot — string len width + wm field name
baremetal_stubs.c: RuntimeString.len uint32_t -> uint64_t. The compiler
inlines text.len() as a 64-bit qword load at offset 8, but the C runtime
stored len as u32 with string data packed right after, so the load pulled
4 bytes of text into the high bits -> a ~quintillion length -> an unbounded
char-emit loop that flooded COM1 and hung boot before the framebuffer
initialized. Widening len to u64 moves data to offset 16 and matches the
compiler/simple-core string ABI.

wm_service.spl: event.mouse_pos -> event.point to match the WmInputEvent
field (the stale name failed HIR type inference and stubbed the module).

+++ examples/simple_os/arch/x86_64/boot/baremetal_stubs.c
@@ -307,7 +307,7 @@ typedef struct {
 
 typedef struct {
     HeapHeader hdr;
-    uint32_t   len;
+    uint64_t   len;
     char       data[];
 } RuntimeString;
